+++ src/components/UI/Icons/constants.ts
@@ -1,0 +1,62 @@
+import {
+  DeleteOutlined, ErrorOutlineOutlined, EditOutlined, LogoutOutlined, Close,
+  DoneOutlined, Visibility, VisibilityOff, KeyboardArrowUpOutlined,
+  Celebration, MonetizationOn,
+  Fastfood, House, Engineering, Language, DirectionsCar,
+  AccountBalance, MedicalInformation, FamilyRestroom, ShoppingCart,
+} from '@mui/icons-material';
+import { AllCategoryIcons, AppIcons } from './interface';
+import { AppColors } from '../../../styles';
+
+export const APP_ICONS: AppIcons = {
+  Delete: {
+    icon: DeleteOutlined,
+    defaultColor: AppColors.negative,
+  },
+  Error: {
+    icon: ErrorOutlineOutlined,
+    defaultColor: AppColors.negative,
+  },
+  Edit: {
+    icon: EditOutlined,
+    defaultColor: AppColors.primary,
+  },
+  LogOut: {
+    icon: LogoutOutlined,
+    defaultColor: AppColors.primary,
+  },
+  Close: {
+    icon: Close,
+    defaultColor: AppColors.grey,
+  },
+  GoToTop: {
+    icon: KeyboardArrowUpOutlined,
+    defaultColor: AppColors.grey,
+  },
+  TickMark: {
+    icon: DoneOutlined,
+    defaultColor: AppColors.positive,
+  },
+  Visibility: {
+    icon: Visibility,
+    defaultColor: AppColors.black,
+  },
+  VisibilityOff: {
+    icon: VisibilityOff,
+    defaultColor: AppColors.black,
+  },
+};
+
+export const categoryIcons: AllCategoryIcons = {
+  foodAndDrink: Fastfood,
+  house: House,
+  utilities: Engineering,
+  subcriptions: Language,
+  transportation: DirectionsCar,
+  financialExpenses: AccountBalance,
+  healthCare: MedicalInformation,
+  kids: FamilyRestroom,
+  shopping: ShoppingCart,
+  entertainment: Celebration,
+  savings: MonetizationOn,
+};