+++ src/components/UI/Records/features/MainRecordDataBox/index.ts
@@ -1,0 +1,1 @@
+export { MainRecordData } from './MainRecordData';