+++ src/components/UI/Records/interface.ts
@@ -10,14 +10,6 @@ export interface RecordProps {
   backgroundColor: string;
 }
 
-export interface DrawerChipContainerProps {
-  afterContent: string;
-}
-
-export interface PaymentStatusChipProps {
-  isPaid?: boolean;
-}
-
 export interface RecordTableProps {
   isGrid?: boolean;
 }