+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## v2.0.1 (2023-12-20)
+## v2.1.0 (2024-03-10)
+Release Changes
+- Change validation for limit of characters for short description and description
+- Add spacing to the theme
+- Change font sizes for mobile
+- Redesign record for all devices
+- Redesign record drawer for all devices
+- Refactor icons component
+- Add icons to category
+- Add limit to validation on adding tags or budgets to 8 maximum
+
+### Pull Requests
+[#3](hhttps://github.com/RafaelMoro/FE_Personal_Finances/pull/3) | Redesign record and record Drawer. Fix other bugs
+
+## v2.0.1 (2024-03-06)
 Release Changes
 - Skip tests to update them using redux. This is to be able to deploy in beta.